Tests

+++ proj1/test/test_backup_2.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 
 TIMEOUT=60
-VERSION=1.0
+VERSION=1.3
 MC_ADDR=230.0.0.1
 MC_PORT=8888
 MDB_ADDR=230.0.0.2
@@ -22,10 +22,10 @@ test () {
         kill $PID5
         exit 1
     fi
-    if [ "$output" != "$expected" ]; then
+    echo $expected > expected.txt
+    echo $output > output.txt
+    if ! diff expected.txt output.txt ; then
         echo -e "\e[1m\e[31m[Failed]\e[0m: expected different from output"
-        echo $expected > expected.txt
-        echo $output > output.txt
         kill $PID1
         kill $PID2
         kill $PID3
@@ -38,7 +38,9 @@ test () {
 
 cd build
 rm -rf 1 2 3 4 5
-curl http://ftp.debian.org/debian/dists/jessie/main/source/Release -o source_Release  # 102B
+if ! [ -f source_Release ]; then curl http://ftp.debian.org/debian/dists/jessie/main/source/Release -o source_Release; fi # 102B
+if ! [ -f Release        ]; then curl http://ftp.debian.org/debian/dists/jessie/Release             -o Release       ; fi # 77.3KB
+if ! [ -f ChangeLog      ]; then curl http://ftp.debian.org/debian/dists/jessie/ChangeLog           -o ChangeLog     ; fi # 2.3MB
 timeout $TIMEOUT java PeerDriver $VERSION 1 service1 $MC_ADDR $MC_PORT $MDB_ADDR $MDB_PORT $MDR_ADDR $MDR_PORT > /dev/null & PID1=$!
 timeout $TIMEOUT java PeerDriver $VERSION 2 service2 $MC_ADDR $MC_PORT $MDB_ADDR $MDB_PORT $MDR_ADDR $MDR_PORT > /dev/null & PID2=$!
 timeout $TIMEOUT java PeerDriver $VERSION 3 service3 $MC_ADDR $MC_PORT $MDB_ADDR $MDB_PORT $MDR_ADDR $MDR_PORT > /dev/null & PID3=$!
@@ -47,11 +49,11 @@ timeout $TIMEOUT java PeerDriver $VERSION 5 service5 $MC_ADDR $MC_PORT $MDB_ADDR
 echo "Started peers with PIDs $PID1, $PID2, $PID3, $PID4, $PID5"
 sleep 1
 
-test "test1-01-1" "timeout $TIMEOUT java TestApp service1 BACKUP source_Release 2" "echo"
-sleep 2
+test "test-backup-2-01-1" "timeout $TIMEOUT java TestApp service1 BACKUP source_Release 2" "echo"
+sleep 10
 FILE=storage/chunks/8C5A4F80497BC0C4719B9DCE7CCC75C36BCB3938A65FB65F7CC0CA0074279526
 contents=$(cat 2/$FILE-0 3/$FILE-0 4/$FILE-0 5/$FILE-0)
-test "test1-01-2" "echo $contents" "cat source_Release source_Release"
+test "test-backup-2-01-2" "echo $contents" "cat source_Release source_Release"
 
 kill $PID1
 kill $PID2
