+++ proj1/test/test_delete_3.sh
@@ -43,10 +43,10 @@ test2 () {
 cd build
 ls
 rm -rf 1 2 3 4
-curl http://ftp.debian.org/debian/dists/jessie/main/source/Release -o source_Release  # 102B
-curl http://ftp.debian.org/debian/dists/jessie/Release             -o Release         # 77.3KB
-curl http://ftp.debian.org/debian/dists/bullseye/InRelease         -o InRelease       # 2.3MB
-timeout $TIMEOUT java PeerDriver $VERSION 1 service1 $MC_ADDR $MC_PORT $MDB_ADDR $MDB_PORT $MDR_ADDR $MDR_PORT  & PID1=$!
+if ! [ -f source_Release ]; then curl http://ftp.debian.org/debian/dists/jessie/main/source/Release -o source_Release; fi # 102B
+if ! [ -f Release        ]; then curl http://ftp.debian.org/debian/dists/jessie/Release             -o Release       ; fi # 77.3KB
+if ! [ -f ChangeLog      ]; then curl http://ftp.debian.org/debian/dists/jessie/ChangeLog           -o ChangeLog     ; fi # 2.3MB
+timeout $TIMEOUT java PeerDriver $VERSION 1 service1 $MC_ADDR $MC_PORT $MDB_ADDR $MDB_PORT $MDR_ADDR $MDR_PORT > /dev/null & PID1=$!
 timeout $TIMEOUT java PeerDriver $VERSION 2 service2 $MC_ADDR $MC_PORT $MDB_ADDR $MDB_PORT $MDR_ADDR $MDR_PORT > /dev/null & PID2=$!
 timeout $TIMEOUT java PeerDriver $VERSION 3 service3 $MC_ADDR $MC_PORT $MDB_ADDR $MDB_PORT $MDR_ADDR $MDR_PORT > /dev/null & PID3=$!
 timeout $TIMEOUT java PeerDriver $VERSION 4 service4 $MC_ADDR $MC_PORT $MDB_ADDR $MDB_PORT $MDR_ADDR $MDR_PORT > /dev/null & PID4=$!
@@ -59,15 +59,15 @@ sleep 3
 
 kill $PID3 # Stops 3
 
-test "test2-01-1" "timeout $TIMEOUT java TestApp service1 DELETE source_Release" "echo"
+test "test-delete-3-01-1" "timeout $TIMEOUT java TestApp service1 DELETE source_Release" "echo"
 sleep 1
 FILE=3/storage/chunks/8C5A4F80497BC0C4719B9DCE7CCC75C36BCB3938A65FB65F7CC0CA0074279526
-test2 "test2-01-2" "$FILE"
+test2 "test-delete-3-01-2" "$FILE"
 
-test "test2-02-1" "timeout $TIMEOUT java TestApp service1 DELETE Release" "echo"
+test "test-delete-3-02-1" "timeout $TIMEOUT java TestApp service1 DELETE Release" "echo"
 sleep 2
 FILE=3/storage/chunks/14C33F2915CA0D86673BCF9A54BC42F73F8A31E0ED6B3EF0D203EAC500F9047D
-test2 "test2-02-2" "$FILE"
+test2 "test-delete-3-02-2" "$FILE"
 
 timeout $TIMEOUT java TestApp service1 DELETE source_Release
 sleep 1
