+++ proj1/test/test_restore_5.sh
@@ -25,13 +25,12 @@ test () {
 
 cd build
 rm -rf 1 2
-rm -rf testfiles
-mkdir testfiles
-curl http://ftp.debian.org/debian/dists/jessie/main/source/Release -o testfiles/source_Release  # 102B
-curl http://ftp.debian.org/debian/dists/jessie/Release             -o testfiles/Release         # 77.3KB
-curl http://ftp.debian.org/debian/dists/jessie/ChangeLog           -o testfiles/ChangeLog       # 2.3MB
-timeout $TIMEOUT java PeerDriver $VERSION 1 service1 $MC_ADDR $MC_PORT $MDB_ADDR $MDB_PORT $MDR_ADDR $MDR_PORT   & PID1=$!
-timeout $TIMEOUT java PeerDriver 1.0 2 service2 $MC_ADDR $MC_PORT $MDB_ADDR $MDB_PORT $MDR_ADDR $MDR_PORT > /dev/null & PID2=$!
+mkdir -p testfiles
+if ! [ -f testfiles/source_Release ]; then curl http://ftp.debian.org/debian/dists/jessie/main/source/Release -o testfiles/source_Release; fi # 102B
+if ! [ -f testfiles/Release        ]; then curl http://ftp.debian.org/debian/dists/jessie/Release             -o testfiles/Release       ; fi # 77.3KB
+if ! [ -f testfiles/ChangeLog      ]; then curl http://ftp.debian.org/debian/dists/jessie/ChangeLog           -o testfiles/ChangeLog     ; fi # 2.3MB
+timeout $TIMEOUT java PeerDriver $VERSION 1 service1 $MC_ADDR $MC_PORT $MDB_ADDR $MDB_PORT $MDR_ADDR $MDR_PORT > /dev/null & PID1=$!
+timeout $TIMEOUT java PeerDriver 1.0      2 service2 $MC_ADDR $MC_PORT $MDB_ADDR $MDB_PORT $MDR_ADDR $MDR_PORT > /dev/null & PID2=$!
 echo "Started peers with PIDs $PID1, $PID2"
 sleep 1
 
@@ -41,7 +40,7 @@ sleep 2
 rm source_Release
 timeout $TIMEOUT java TestApp service1 RESTORE source_Release
 sleep 2
-test "test-restore-1-01" "cat source_Release" "cat testfiles/source_Release"
+test "test-restore-5-01" "cat source_Release" "cat testfiles/source_Release"
 
 cp testfiles/Release .
 timeout $TIMEOUT java TestApp service1 BACKUP Release 1
@@ -49,7 +48,7 @@ sleep 3
 rm Release
 timeout $TIMEOUT java TestApp service1 RESTORE Release
 sleep 4
-test "test-restore-1-02" "cat Release" "cat testfiles/Release"
+test "test-restore-5-02" "cat Release" "cat testfiles/Release"
 
 cp testfiles/ChangeLog .
 timeout $TIMEOUT java TestApp service1 BACKUP ChangeLog 1
@@ -57,7 +56,7 @@ sleep 40
 rm ChangeLog
 timeout $TIMEOUT java TestApp service1 RESTORE ChangeLog
 sleep 30
-test "test-restore-1-03" "cat ChangeLog" "cat testfiles/ChangeLog"
+test "test-restore-5-03" "cat ChangeLog" "cat testfiles/ChangeLog"
 
 kill $PID1
 kill $PID2
